pyproject.toml
[tool.poetry]
name = "kbe"
version = "0.1.0"
description = ""
authors = ["julitschkaa <schaak@posteo.net>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
